pyproject.toml
[project]
name = "unify-llm"
version = "0.1.0"
description = "统一大语言模型框架"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "anthropic>=0.69.0",
    "datasets>=4.2.0",
    "google-genai>=1.43.0",
    "numpy>=2.2.6",
    "openai>=2.3.0",
    "pydantic>=2.12.1",
    "pyyaml>=6.0.3",
    "redis>=6.4.0",
    "scikit-learn>=1.7.2",
    "torch>=2.8.0",
    "transformers>=4.57.0",
    "vllm>=0.11.0",
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
src_paths = ["unify_llm", "tests"]

[tool.mypy]
strict = true
ignore_missing_imports = true
python_version = "3.12"
show_error_codes = true
disable_error_code = []

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
]

[tool.setuptools]
packages = ["unify_llm"]

[tool.uv]
index-url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
